jqftu: Initial jqftu module
'jqftu' is a Japanese quiz module. It works just like the Kotoba bot on
Discord. Currently, the available deck is only Tozai Line. I'll add more
decks later. I hope given this module, we can memorize Japanese kanji
and words faster.

Additionally, add json module from nlohmann. It's needed to run the
jqftu module. Note that the json module is not private to the jqftu
module, so other modules can also use it in the future.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>